bug : recipe list component doesn't print the right state on mount
Contexte : when editing recipe from database, the steps are note correctly printed althoug the value of the step is correct.State is correctly printed after on click. Tried to simulate on click by calling the function after mounting element -> value changed but not the component apperence
