+++ src/js/frame/Home.js
@@ -6,6 +6,7 @@ import Grafana from '../../Grafana'
 import logo from '../../zba.svg';
 import axios from 'axios';
 import '../../css/App.css'
+import Step from '../recipe/Step'
 
 class Home extends Component {
     constructor() {
@@ -110,14 +111,31 @@ class Home extends Component {
                     creator: creator
                 }
             })
-            if(id != '') {
+            if(id !== '') {
                 axios.get(`http://localhost:8080/Steps${id}`)
                 .then(result => {
-                    console.log(result.data)
-                    const res = result.data
-                    this.setState({ recipeSteps: {steps: res} }, () => 
+                    let steps = this.state.recipeSteps.steps
+                    let id = this.state.recipeSteps.nextStepId
+                    result.data.forEach((element) => {
+                        steps.push({
+                            component: <Step
+                                x={this.handleXClick}
+                                getValue={this.getValue}
+                                handleOnChange={this.handleOnChange}
+                                id={element.id}
+                            />,
+                            selectedStep: element.selectedStep,
+                            description: element.description,
+                            id: element.id
+                        })
+                        this.handleOnChange(element.id, 2, "selectedStep")
+                        console.log(this.getValue(this.state.recipeSteps.steps[0].id,"selectedStep"))
+                    })
+                    this.setState({ recipeSteps: {steps: steps} }, () =>
                         this.setState(prevState => ({control:{editMode: true}}))
-                    );
+                    )
+                    id++
+                    this.setListId(id)
                     this.setUpdate()
                 })
             } else {
@@ -129,13 +147,65 @@ class Home extends Component {
     setListSteps = (newSteps) => {
         let object = this.state.recipeSteps
         object.steps = newSteps
-        this.setState({recipeSteps: object})
+        this.setState({recipeSteps: object}, () => console.log(this.state.recipeSteps))
     }
 
     setListId = (newId) => {
         let object = this.state.recipeSteps
         object.nextStepId = newId
         this.setState({recipeSteps: object})
+    }
+
+    getValue = (id, subElement) => {
+        let details
+        this.state.recipeSteps.steps.forEach((element) => {
+            if(element.id === id) {
+                details = element[subElement]
+            }
+        })
+        return details
+    }
+
+    handleOnChange = (id, value, subElement) => {
+        const array = this.state.recipeSteps.steps.map((element) => {
+            if(element.id === id) {
+                const newElement = element
+                newElement[subElement] = value
+                return newElement
+            } else {
+                return element
+            }
+        })
+        this.setListSteps(array)
+    }
+
+    handleXClick = (id) => {
+        var array = []
+        this.state.recipeSteps.steps.forEach((element) => {
+            if(element.id !== id) {
+                array = array.concat(element)
+            }
+        })
+        this.setListSteps(array)
+    }
+
+    handleNewClick = () => {
+        let steps = this.state.recipeSteps.steps
+        let id = this.state.recipeSteps.nextStepId
+        steps.push({
+            component: <Step
+                x={this.handleXClick}
+                getValue={this.getValue}
+                handleOnChange={this.handleOnChange}
+                id={id}
+            />,
+            selectedStep: "1",
+            description: "",
+            id: id
+        })
+        id++
+        this.setListId(id)
+        this.setListSteps(steps)
     }
 
     render() {
@@ -149,7 +219,7 @@ class Home extends Component {
                     {this.state.control.showRecipeTable ?
                         (<RecipeTable setEdit={this.setEdit} update={this.state.update} setUpdate={this.setUpdate}/>)
                         :
-                        (<RecipeStep steps={this.state.recipeSteps.steps} setListSteps={this.setListSteps} id={this.state.recipeSteps.nextStepId} setListId={this.setListId} />)
+                        (<RecipeStep steps={this.state.recipeSteps.steps} setListSteps={this.setListSteps} />)
                     }
                     <Grafana/>
                 </div>
