+++ src/js/recipe/RecipeStep.js
@@ -1,11 +1,9 @@
 import React, { Component } from 'react';
 import {SortableContainer, SortableElement} from 'react-sortable-hoc';
 import arrayMove from 'array-move';
-import Step from './Step'
 import "../../css/App.css"
 
 const SortableItem = SortableElement(({value}) => {
-    console.log(value)
     return <div>{value.component}</div>
 })
 
@@ -20,58 +18,6 @@ const SortableList = SortableContainer(({steps}) => {
 })
 
 class RecipeStep extends Component {
-
-    getValue = (id, subElement) => {
-        let details
-        this.props.steps.forEach((element) => {
-            if(element.id === id) {
-                details = element[subElement]
-            }
-        })
-        return details
-    }
-
-    handleOnChange = (id, value, subElement) => {
-        const array = this.props.steps.map((element) => {
-            if(element.id === id) {
-                const newElement = element
-                newElement[subElement] = value
-                return newElement
-            } else {
-                return element
-            }
-        })
-        this.props.setListSteps(array)
-    }
-
-    handleXClick = (id) => {
-        var array = []
-        this.props.steps.forEach((element) => {
-            if(element.id !== id) {
-                array = array.concat(element)
-            }
-        })
-        this.props.setListSteps(array)
-    }
-
-    handleNewClick = () => {
-        let steps = this.props.steps
-        let id = this.props.id
-        steps.push({
-            component: <Step
-                x={this.handleXClick}
-                getValue={this.getValue}
-                handleOnChange={this.handleOnChange}
-                id={id}
-            />,
-            selectedStep: "1",
-            description: "",
-            id: id
-        })
-        id++
-        this.props.setListId(id)
-        this.props.setListSteps(steps)
-    }
 
     onSortEnd = ({oldIndex, newIndex}) => {
         let steps = this.props.steps
